On wrong password, show popup and delete saved pw
Fixes #200
